+++ Q3.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Sun Dec 22 16:19:24 2019
+
+@author: joyce
+"""
+
+from variables import cenarios2
